test(Track...ReroutingTest): Split rerouting code from TrackedTripControllerTest.
